requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
autopep8==1.5.7
bleach==3.3.0
build==0.5.1
certifi==2021.5.30
cffi==1.14.5
chardet==4.0.0
colorama==0.4.4
cycler==0.10.0
discord.py==1.7.3
dnspython==2.1.0
PyNaCl==1.4.0